refactor: simplify demo event log rendering
